Separated out logic for lenv, lval, and builtin
