+++ builtin.h
@@ -1,0 +1,21 @@
+#ifndef builtin_h
+#define builtin_h
+
+#include "lval.h"
+#include "lenv.h"
+
+lval* builtin_op(lenv* e, lval* a, char* op);
+lval* builtin_add(lenv* e, lval* a);
+lval* builtin_sub(lenv* e, lval* a);
+lval* builtin_mul(lenv* e, lval* a);
+lval* builtin_div(lenv* e, lval* a);
+lval* builtin_head(lenv* e, lval* a);
+lval* builtin_tail(lenv* e, lval* a);
+lval* builtin_list(lenv* e, lval* a);
+lval* builtin_eval(lenv* e, lval* a);
+lval* builtin_join(lenv* e, lval* a);
+lval* builtin_def(lenv* e, lval* a);
+lval* builtin_len(lenv* e, lval* a);
+lval* builtin(lenv* e, lval* a, char* func);
+
+#endif
