+++ rok.c
@@ -2,6 +2,9 @@
 #include <stdlib.h>
 #include <math.h>
 #include "mpc.h"
+#include "builtin.h"
+#include "lenv.h"
+#include "lval.h"
 
 /* If we are compiling on Windows compile these functions */
 #ifdef _WIN32
@@ -27,144 +30,6 @@ void add_history(char* unused) {}
 #include <editline/readline.h>
 #endif
 
-/* Forward Declarations */
-struct lval;
-struct lenv;
-typedef struct lval lval;
-typedef struct lenv lenv;
-void lval_print(lval* v);
-lval* lval_eval(lenv* e, lval* v);
-void lenv_put(lenv* e, lval* v, lval* k);
-
-/* Declare Enumerations for lval types */
-enum lval_types { LVAL_NUM, LVAL_ERR, LVAL_SYM, LVAL_SEXPR, LVAL_QEXPR, LVAL_FUN };
-
-typedef lval*(*lbuiltin)(lenv*, lval*);
-
-/* Declare new lval struct */
-struct lval {
-  int type;
-
-  /* Basic */
-  long num;
-  char* err;
-  char* sym;
-
-  /* Function */
-  lbuiltin builtin;
-  lenv* env;
-  lval* formals;
-  lval* body;
-
-  /* Expression */
-  int count;
-  lval** cell;
-};
-
-struct lenv {
-  int count;
-  char** syms;
-  lval** vals;
-};
-
-/** Lval functions **/
-lval* lval_num(long x) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_NUM;
-  v->num = x;
-  return v;
-}
-
-lval* lval_err(char* fmt, ...) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_ERR;
-
-  /* Create a va_list and initialize it */
-  va_list va;
-  va_start(va, fmt);
-
-  v->err = malloc(512);
-
-  /* Print the error string with a maximum of 511 characters */
-  vsnprintf(v->err, 511, fmt, va);
-
-  /* Reallocate to number of bytes actually used */
-  v->err = realloc(v->err, strlen(v->err)+1);
-
-  /* Clean up VA list */
-  va_end(va);
-  return v;
-}
-
-lval* lval_sym(char* sym) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_SYM;
-  v->sym = malloc(strlen(sym) + 1);
-  strcpy(v->sym, sym);
-  return v;
-}
-
-lval* lval_sexpr(void) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_SEXPR;
-  v->count = 0;
-  v->cell = NULL;
-  return v;
-}
-
-lval* lval_qexpr(void) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_QEXPR;
-  v->count = 0;
-  v->cell = NULL;
-  return v;
-}
-
-lval* lval_fun(lbuiltin builtin) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_FUN;
-  v->builtin = builtin;
-  return v;
-}
-
-lval* lval_add(lval* v, lval* x) {
-  v->count++;
-  v->cell = realloc(v->cell, sizeof(lval*) * v->count);
-  v->cell[v->count-1] = x;
-  return v;
-}
-
-void lval_del(lval* v) {
-  switch (v->type) {
-    /* Do nothing special for number type */
-    case LVAL_NUM: break;
-
-    /* For err or sym free the data */
-    case LVAL_ERR: free(v->err); break;
-    case LVAL_SYM: free(v->sym); break;
-    case LVAL_FUN:
-      if (!v->builtin) {
-        lenv_del(v->env);
-        lval_del(v->formals);
-        lval_del(v->body);
-      }
-    break;
-
-    /* If qexpr or sexpr then delete all elements inside */
-    case LVAL_QEXPR:
-    case LVAL_SEXPR:
-      for (int i = 0; i < v->count; i++) {
-        lval_del(v->cell[i]);
-      }
-
-      /* Free memory allocated to store pointers */
-      free(v->cell);
-    break;
-  }
-  /* Free memory allocated for the lval struct itself */
-  free(v);
-}
-
 char* ltype_name(int t) {
   switch(t) {
     case LVAL_FUN: return "Function";
@@ -183,473 +48,6 @@ char* ltype_name(int t) {
     lval_del(args); \
     return err; \
   }
-
-lval* lval_read_num(mpc_ast_t* t) {
-  errno = 0;
-  long x = strtol(t->contents, NULL, 10);
-  return errno != ERANGE ?
-    lval_num(x) : lval_err("invalid number");
-}
-
-lval* lval_read(mpc_ast_t* t) {
-  /* If symbol or number return conversion to that type */
-  if (strstr(t->tag, "number")) { return lval_read_num(t); }
-  if (strstr(t->tag, "symbol")) { return lval_sym(t->contents); }
-
-  /* If root (>) or sexpr then create empty list */
-  lval* x = NULL;
-  if (strcmp(t->tag, ">") == 0) { x = lval_sexpr(); }
-  if (strstr(t->tag, "sexpr"))  { x = lval_sexpr(); }
-  if (strstr(t->tag, "qexpr"))  { x = lval_qexpr(); }
-
-  /* Fill this list with any valid expression contained within */
-  for (int i = 0; i < t->children_num; i++) {
-    if (strcmp(t->children[i]->contents, "(") == 0) { continue; }
-    if (strcmp(t->children[i]->contents, ")") == 0) { continue; }
-    if (strcmp(t->children[i]->contents, "{") == 0) { continue; }
-    if (strcmp(t->children[i]->contents, "}") == 0) { continue; }
-    if (strcmp(t->children[i]->tag,  "regex") == 0) { continue; }
-    x = lval_add(x, lval_read(t->children[i]));
-  }
-
-  return x;
-}
-
-
-void lval_expr_print(lval* v, char open, char close) {
-  putchar(open);
-  for (int i = 0; i < v->count; i++) {
-    /* Print value contained within */
-    lval_print(v->cell[i]);
-
-    /* Don't print trailing space if last element */
-    if (i != (v->count-1)) {
-      putchar(' ');
-    }
-  }
-  putchar(close);
-}
-
-void lval_print(lval* v) {
-  switch(v->type) {
-    /* In the case the type is a number print it */
-    /* Then break out of the switch */
-    case LVAL_NUM: printf("%li", v->num); break;
-    case LVAL_ERR: printf("Error: %s", v->err); break;
-    case LVAL_SYM: printf("%s", v->sym); break;
-    case LVAL_FUN:
-      if (v->builtin) {
-        printf("<builtin>");
-      } else {
-        printf("(\\ "); lval_print(v->formals);
-        putchar(' '); lval_print(v->body); putchar(')');
-      }
-    break;
-    case LVAL_SEXPR: lval_expr_print(v, '(', ')'); break;
-    case LVAL_QEXPR: lval_expr_print(v, '{', '}'); break;
-    break;
-  }
-}
-
-lval* lval_pop(lval* v, int i) {
-  /* Find the item at i */
-  lval* x = v->cell[i];
-
-  /* Shift memory after the item at "i" over the top */
-  /* WTF is memmove? */
-  memmove(&v->cell[i], &v->cell[i+1],
-    sizeof(lval*) * (v->count-i-1));
-
-  /* Decrease the count of items in the list */
-  v->count--;
-
-  /* Reallocate the memory used */
-  v->cell = realloc(v->cell, sizeof(lval*) * v->count);
-  return x;
-}
-
-lval* lval_take(lval* v, int i) {
-  lval* x = lval_pop(v, i);
-  lval_del(v);
-  return x;
-}
-
-
-lval* lval_join(lval* x, lval* y) {
-  /* For each cell in 'y' add it to 'x' */
-  while(y->count) {
-    x = lval_add(x, lval_pop(y, 0));
-  }
-
-  /* Delete the empty y and return x */
-  lval_del(y);
-  return x;
-}
-
-lval* lval_copy(lval* v) {
-  lval* x = malloc(sizeof(lval));
-  x->type = v->type;
-
-  switch (v->type) {
-    case LVAL_NUM: x->num = v->num; break;
-    case LVAL_FUN:
-      if (v->builtin) {
-        x->builtin = v->builtin;
-      } else {
-        x->builtin = NULL;
-        x->env = lenv_copy(v->env);
-        x->formals = lval_copy(v->formals);
-        x->body = lval_copy(v->body);
-      }
-    break;
-    case LVAL_SYM:
-      x->sym = malloc(strlen(v->sym) + 1);
-      strcpy(x->sym, v->sym); break;
-    case LVAL_ERR:
-      x->err = malloc(strlen(v->err) + 1);
-      strcpy(x->err, v->err); break;
-
-    case LVAL_SEXPR:
-    case LVAL_QEXPR:
-      x->count = v->count;
-      x->cell = malloc(sizeof(lval*) * x->count);
-      for (int i = 0; i < x->count; i++) {
-        x->cell[i] = lval_copy(v->cell[i]);
-      }
-    break;
-  }
-
-  return x;
-}
-
-lval* lval_lambda(lval* formals, lval* body) {
-  lval* v = malloc(sizeof(lval));
-  v->type = LVAL_FUN;
-
-  /* Set builtin to null */
-  v->builtin = NULL;
-
-  /* build new environment */
-  v->env = lenv_new();
-
-  /* set formals and body */
-  v->formals = formals;
-  v->body = body;
-  return v;
-}
-
-lval* builtin_op(lenv* e, lval* a, char* op) {
-  /* Ensure all arguments are numbers */
-  for (int i = 0; i < a->count; i++) {
-    if(a->cell[i]->type != LVAL_NUM) {
-      lval_del(a);
-      return lval_err("Cannot operate on non-numbers");
-    }
-  }
-
-  /* Pop the first element */
-  lval* x = lval_pop(a, 0);
-
-  /* If no arguments and sub then perform unary negation */
-  if ((strcmp(op, "-") == 0) && a->count == 0) {
-    x->num = -x->num;
-  }
-
-  /* While there are still elements remaining */
-  while (a->count > 0) {
-    /* Pop the next element */
-    lval* y = lval_pop(a, 0);
-
-    if(strcmp(op, "+") == 0) { x->num += y->num; }
-    if(strcmp(op, "-") == 0) { x->num -= y->num; }
-    if(strcmp(op, "*") == 0) { x->num *= y->num; }
-    if(strcmp(op, "/") == 0) {
-      if(y->num == 0) {
-        lval_del(x); lval_del(y);
-        x = lval_err("Division By Zero!"); break;
-      }
-      x->num /= y->num;
-    }
-    if(strcmp(op, "%") == 0) { x->num %= y->num; }
-    if(strcmp(op, "^") == 0) { x->num = (long)pow(x->num, y->num);}
-    lval_del(y);
-  }
-
-  lval_del(a); return x;
-}
-
-lval* builtin_add(lenv* e, lval* a) {
-  return builtin_op(e, a, "+");
-}
-
-lval* builtin_sub(lenv* e, lval* a) {
-  return builtin_op(e, a, "-");
-}
-
-lval* builtin_mul(lenv* e, lval* a) {
-  return builtin_op(e, a, "*");
-}
-
-lval* builtin_div(lenv* e, lval* a) {
-  return builtin_op(e, a, "/");
-}
-
-lval* builtin_head(lenv* e, lval* a) {
-  /* Check Error Conditions */
-  int count = 1;
-  LASSERT(a, a->cell[0]->type == LVAL_QEXPR,
-     "Function 'head' passed incorrect types! \n"
-     "Got %s, Expected %s.",
-     ltype_name(a->cell[0]->type), ltype_name(LVAL_QEXPR));
-  LASSERT(a, a->cell[0]->count != 0,
-   "Function 'head' passed {}!");
-  LASSERT(a, a->count == count,
-    "Function 'head' passed too many arguments! \n"
-    "Got %i, Expected %i.",
-    a->count, 1);
-
-  /* Otherwise take first argument */
-  lval* v = lval_take(a, 0);
-
-  /* Delete all elements that are not head and return */
-  while(v->count > 1) { lval_del(lval_pop(v, 1)); }
-  return v;
-}
-
-lval* builtin_tail(lenv* e, lval* a) {
-  /* Check Error Conditions */
-  LASSERT(a, a->count == 1,
-    "Function 'tail' passed too many arguments! \n"
-    "Got %i, Expected %i",
-    a->count, 1);
-  LASSERT(a, a->cell[0]->type == LVAL_QEXPR,
-    "Function 'tail' passed incorrect types! \n"
-    "Got %s, Expected %s",
-    ltype_name(a->cell[0]->type), ltype_name(LVAL_QEXPR));
-  LASSERT(a, a->cell[0]->count != 0,
-    "Function 'tail' passed {}!");
-
-  /* Other wise take first arg */
-  lval* v = lval_take(a, 0);
-
-  /* Delete first element and return */
-  lval_del(lval_pop(v, 0));
-  return v;
-}
-
-lval* builtin_list(lenv* e, lval* a) {
-  a->type = LVAL_QEXPR;
-  return a;
-}
-
-lval* builtin_eval(lenv* e, lval* a) {
-  LASSERT(a, a->count == 1,
-    "Function 'eval' passed too many arguments! \n"
-    "Got %s, Expected %s",
-    a->count, 1);
-  LASSERT(a, a->cell[0]->type == LVAL_QEXPR,
-     "Function 'eval' passed incorrect type! \n"
-     "Got %s, Expected %s",
-     ltype_name(a->cell[0]->type), ltype_name(LVAL_QEXPR));
-  lval* x = lval_take(a, 0);
-  x->type = LVAL_SEXPR;
-  return lval_eval(e, x);
-}
-
-lval* builtin_join(lenv* e, lval* a) {
-  for(int i = 0; i < a->count; i++) {
-    LASSERT(a, a->cell[0]->type == LVAL_QEXPR,
-      "Function 'join' passed incorrect types! \n"
-      "Got %s, Expected %s",
-      ltype_name(a->cell[0]->type), ltype_name(LVAL_QEXPR));
-  }
-
-  lval* x = lval_pop(a, 0);
-
-  while(a->count) {
-    x = lval_join(x, lval_pop(a, 0));
-  }
-
-  lval_del(a);
-  return x;
-}
-
-lval* builtin_def(lenv* e, lval* a) {
-  LASSERT(a, a->cell[0]->type == LVAL_QEXPR,
-    "Function 'def' passed incorrect type! \n"
-    "Got %s, Expected %s",
-    ltype_name(a->cell[0]->type), ltype_name(LVAL_QEXPR));
-
-  /* First argument is symbol list */
-  lval* syms = a->cell[0];
-
-  /* Ensure all elements of first list are symbols */
-  for(int i = 0; i < syms->count; i++) {
-    LASSERT(a, syms->cell[i]->type == LVAL_SYM,
-    "Function 'def' cannot define non-symbols! \n"
-    "Got %s, Expected %s",
-    ltype_name(syms->cell[i]->type), ltype_name(LVAL_SYM));
-  }
-
-  /* Check correct number of symbols and values */
-    LASSERT(a, syms->count == a->count-1,
-    "Function 'def' cannot define incorrect "
-    "number of values to symbols. \n"
-    "Got %i values, Expected %i values",
-    syms->count, a->count-1);
-
-  /* Assign copies of values to symbols */
-  for (int i = 0; i < syms->count; i++) {
-    lenv_put(e, syms->cell[i], a->cell[i+1]);
-  }
-
-  lval_del(a);
-  return lval_sexpr();
-}
-
-lval* builtin_cons(lenv* e, int val, lval* a) {
-  return a;
-}
-
-// Return the number of elements in a Qexpression
-lval* builtin_len(lenv* e, lval* a) {
-  return lval_num(a->cell[0]->count);
-}
-
-lval* builtin(lenv* e, lval* a, char* func) {
-  if(strcmp("list", func) == 0) { return builtin_list(e, a); }
-  if(strcmp("head", func) == 0) { return builtin_head(e, a); }
-  if(strcmp("tail", func) == 0) { return builtin_tail(e, a); }
-  if(strcmp("eval", func) == 0) { return builtin_eval(e, a); }
-  if(strcmp("join", func) == 0) { return builtin_join(e, a); }
-  if(strcmp("len", func) == 0) { return builtin_len(e, a); }
-  if(strcmp("def", func) == 0) { return builtin_def(e, a); }
-  if(strstr("+-/*^%", func)) { return builtin_op(e, a, func); }
-  lval_del(a);
-  return lval_err("Unknown Function!");
-}
-
-lval* lval_eval_sexpr(lenv* e, lval* v) {
-
-  /* Evaluate Children */
-  for (int i = 0; i < v->count; i++) {
-    v->cell[i] = lval_eval(e, v->cell[i]);
-  }
-
-  /* Error checking */
-  for (int i = 0; i < v->count; i++) {
-    if (v->cell[i]->type == LVAL_ERR) { return lval_take(v, i); }
-  }
-
-  /* Empty Expression */
-  if (v->count == 0) { return v; }
-
-  /* Single Expression */
-  if (v->count == 1) { return lval_take(v, 0); }
-
-  /* Ensure First Element is a function after evaluation */
-  lval* f = lval_pop(v, 0);
-  if (f->type != LVAL_FUN) {
-    lval_del(f); lval_del(v);
-    return lval_err("'%s' is not a function!", f->type);
-  }
-
-  /* If so call function to get result */
-  lval* result = f->builtin(e, v);
-  lval_del(f);
-  return result;
-}
-
-lval* lenv_get(lenv* e, lval* k) {
-  /* Iterate over all items in the environment */
-  for (int i = 0; i < e->count; i++) {
-    if(strcmp(e->syms[i], k->sym) == 0) {
-      return lval_copy(e->vals[i]);
-    }
-  }
-  return lval_err("unbound symbol '%s'!", k->sym);
-}
-
-void lenv_put(lenv* e, lval* k, lval* v) {
-  /* Iterate */
-  for (int i = 0; i < e->count; i++) {
-    if(strcmp(e->syms[i], k->sym) == 0) {
-      lval_del(e->vals[i]);
-      e->vals[i] = lval_copy(v);
-      return;
-    }
-  }
-
-  /* If no existing entry found, allocate space for new entry */
-  e->count++;
-  e->vals = realloc(e->vals, sizeof(lval*) * e->count);
-  e->syms = realloc(e->syms, sizeof(char*) * e->count);
-
-  /* Copy contents of lval and symbol string to new location */
-  e->vals[e->count-1] = lval_copy(v);
-  e->syms[e->count-1] = malloc(strlen(k->sym)+1);
-  strcpy(e->syms[e->count-1], k->sym);
-}
-
-lval* lval_eval(lenv* e, lval* v) {
-  if (v->type == LVAL_SYM) {
-    lval* x = lenv_get(e, v);
-    lval_del(v);
-    return x;
-  }
-  /* Evaluate Sexpressions */
-  if (v->type == LVAL_SEXPR) { return lval_eval_sexpr(e, v); }
-  /* All other lval types remain the same */
-  return v;
-}
-
-/* Print an "lval" followed by a newline */
-void lval_println(lval* v) { lval_print(v); putchar('\n'); }
-
-/** Lenv functions **/
-lenv* lenv_new(void) {
-  lenv* e = malloc(sizeof(lenv));
-  e->count = 0;
-  e->syms = NULL;
-  e->vals = NULL;
-  return e;
-}
-
-void lenv_del(lenv* e) {
-  for(int i = 0; i < e->count; i++) {
-    free(e->syms[i]);
-    lval_del(e->vals[i]);
-  }
-  free(e->syms);
-  free(e->vals);
-  free(e);
-}
-
-void lenv_add_builtin(lenv* e, char* name, lbuiltin func) {
-  lval* k = lval_sym(name);
-  lval* v = lval_fun(func);
-  lenv_put(e, k, v);
-  lval_del(k); lval_del(v);
-}
-
-void lenv_add_builtins(lenv* e) {
-  /* List functions */
-  lenv_add_builtin(e, "list", builtin_list);
-  lenv_add_builtin(e, "head", builtin_head);
-  lenv_add_builtin(e, "tail", builtin_tail);
-  lenv_add_builtin(e, "eval", builtin_eval);
-  lenv_add_builtin(e, "len", builtin_len);
-  lenv_add_builtin(e, "join", builtin_join);
-
-  /* Math functions */
-  lenv_add_builtin(e, "+", builtin_add);
-  lenv_add_builtin(e, "-", builtin_sub);
-  lenv_add_builtin(e, "*", builtin_mul);
-  lenv_add_builtin(e, "/", builtin_div);
-
-  /* Variable functions */
-  lenv_add_builtin(e, "def", builtin_def);
-}
 
 int main(int argc, char** argv) {
   /* Create some parsers */
